create animated count in header
