+++ src/partials/hero-header.html
@@ -162,7 +162,7 @@
       <img src="images/img-for-hero-header-desktop/tasty.png" alt="Delighted girl" />
     </picture>
 
-    <picture class="hero__cafes img-pos hero__image--hide animate__animated animate__swing">
+    <!-- <picture class="hero__cafes img-pos hero__image--hide animate__animated animate__swing">
       <source
         media="(min-width: 1280px)"
         srcset="
@@ -200,7 +200,12 @@
         sizes="43px"
       />
       <img src="images/img-for-hero-header-desktop/foodtrucks.png" alt="Number of foodtrucks" />
-    </picture>
+    </picture> -->
+
+    <div class="hero__count-img hero__cafes animate__animated animate__swing"><span id="out-1" class="hero__count-text"></span><span class="hero__count-type"><pre>      cafes</pre></span>
+    </div>
+    <div class="hero__count-img hero__foodtracks animate__animated animate__swing"><span id="out-2" class="hero__count-text"></span><span class="hero__coint-type"><pre>food trucks    </pre></span>
+    </div>
   </div>
 </section>
