+++ src/sass/base/_reset.scss
@@ -9,6 +9,13 @@ a {
   font-style: normal;
 }
 
+pre {
+  font-family: inherit;
+  margin: 0;
+  text-decoration: none;
+  font-style: normal;
+}
+
 h1,
 h2,
 h3,
