+++ src/sass/components/header+hero/_header.scss
@@ -9,6 +9,11 @@
       justify-content: space-between;
       padding-top: 6px;
     }
+
+    @media screen and (min-width: 480px) and (max-width: 767px) {
+      width: 100vw;
+    }
+
     @media screen and (min-width: 768px) {
       justify-content: start;
       align-items: center;
