+++ src/sass/components/header+hero/_hero.scss
@@ -7,5 +7,9 @@
     position: relative;
     display: flex;
     flex-direction: column;
+
+    @media screen and (min-width: 480px) and (max-width: 767px) {
+      width: 100vw;
+    }
   }
 }
