+++ src/sass/utils/_variables.scss
@@ -1,4 +1,4 @@
-$color-main: #D41342; // Лого, заголовки
+$color-main: #d41342; // Лого, заголовки
 $color-second: #ffa5ba; // Текст над заголовком
 $color-light: #fff; // Белый текст
 $color-midle: #907e82; // Серый текст
